requirements.txt
streamlit==1.41.1
pandas==2.2.3
langchain==0.3.17
langchain-groq==0.2.4
python-dotenv==1.0.1
langchain-community==0.3.16